+++ Apps/Contoso.Forms.Puppet/Contoso.Forms.Puppet/ModulePages/AppCenterContentPage.xaml.cs
@@ -33,6 +33,7 @@ namespace Contoso.Forms.Puppet
         {
             base.OnAppearing();
             AppCenterEnabledSwitchCell.On = await AppCenter.IsEnabledAsync();
+            AllowedNetworkRequestSwitchCell.On = AppCenter.IsNetworkRequestsAllowed;
             if (Application.Current.Properties.ContainsKey(Constants.UserId) && Application.Current.Properties[Constants.UserId] is string id)
             {
                 UserIdEntry.Text = id;
@@ -86,6 +87,11 @@ namespace Contoso.Forms.Puppet
             {
                 AppCenterLog.Error(LogTag, "Wrong number value for the max storage size.");
             }
+        }
+
+        void AllowedNetworkRequestEnabled(System.Object sender, Xamarin.Forms.ToggledEventArgs e)
+        {
+            AppCenter.IsNetworkRequestsAllowed = e.Value;
         }
     }
 }