+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Channel/Channel.cs
@@ -275,14 +275,22 @@ namespace Microsoft.AppCenter.Channel
             }
         }
 
-        private void Resume(State state)
+        /// <summary>
+        /// Resume channel.
+        /// </summary>
+        /// <param name="state">Current state.</param>
+        /// <param name="needEnableChannel">Value indicating whether channel should be enabled. True by default.</param>
+        private void Resume(State state, bool needEnableChannel = true)
         {
             AppCenterLog.Debug(AppCenterLog.LogTag, $"Resume channel: '{Name}'");
             try
             {
                 using (_mutex.GetLock(state))
                 {
-                    _enabled = true;
+                    if (needEnableChannel)
+                    {
+                        _enabled = true;
+                    }
                     _discardLogs = false;
                     state = _mutex.InvalidateState();
                 }
@@ -294,7 +302,14 @@ namespace Microsoft.AppCenter.Channel
             CheckPendingLogs(state);
         }
 
-        private void Suspend(State state, bool deleteLogs, Exception exception)
+        /// <summary>
+        /// Suspend channel.
+        /// </summary>
+        /// <param name="state">Current state.</param>
+        /// <param name="deleteLogs">Value indicating whether logs should be enabled or disabled.</param>
+        /// <param name="exception">Possible error if unsuccessful.</param>
+        /// <param name="needDisableChannel">Value indicating whether channel should be disabled. True by default.</param>
+        private void Suspend(State state, bool deleteLogs, Exception exception, bool needDisableChannel = true)
         {
             AppCenterLog.Debug(AppCenterLog.LogTag, $"Suspend channel: '{Name}'");
             try
@@ -303,7 +318,10 @@ namespace Microsoft.AppCenter.Channel
                 IList<Log> unsentLogs = null;
                 using (_mutex.GetLock(state))
                 {
-                    _enabled = false;
+                    if (needDisableChannel)
+                    {
+                        _enabled = false;
+                    }
                     _batchScheduled = false;
                     _discardLogs = deleteLogs;
                     if (deleteLogs)
@@ -555,6 +573,18 @@ namespace Microsoft.AppCenter.Channel
             }
         }
 
+        public void SetNetworkRequestAllowed(bool isAllowed)
+        {
+            if (isAllowed)
+            {
+                Resume(_mutex.State, false);
+            }
+            else
+            {
+                Suspend(_mutex.State, false, new CancellationException(), false);
+            }
+        }
+
         private void CheckPendingLogs(State state)
         {
             if (!_enabled)
@@ -562,7 +592,11 @@ namespace Microsoft.AppCenter.Channel
                 AppCenterLog.Info(AppCenterLog.LogTag, "The service has been disabled. Stop processing logs.");
                 return;
             }
-
+            if (!_ingestion.IsEnabled)
+            {
+                AppCenterLog.Info(AppCenterLog.LogTag, "App Center is in offline mode.");
+                return;
+            }
             AppCenterLog.Debug(AppCenterLog.LogTag, $"CheckPendingLogs({Name}) pending log count: {_pendingLogCount}");
             using (_mutex.GetLock())
             {