+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Channel/IChannel.cs
@@ -46,5 +46,11 @@ namespace Microsoft.AppCenter.Channel
         /// Stop all calls in progress and deactivate this channel
         /// </summary>
         Task ShutdownAsync();
+
+        /// <summary>
+        /// Set network request allowed. If true check pending logs, suspend sending logs otherwise.
+        /// </summary>
+        /// <param name="isAllowed">True if network request allowed, false otherwise.</param>
+        void SetNetworkRequestAllowed(bool isAllowed);
     }
 }