+++ SDK/AppCenter/Microsoft.AppCenter.Windows.Shared/Ingestion/Http/IngestionDecorator.cs
@@ -35,5 +35,10 @@ namespace Microsoft.AppCenter.Ingestion.Http
         {
             DecoratedApi.Dispose();
         }
+
+        public bool IsEnabled
+        {
+            get => DecoratedApi.IsEnabled;
+        }
     }
 }