+++ SDK/AppCenter/Microsoft.AppCenter.iOS.Bindings/ApiDefinition.cs
@@ -235,6 +235,16 @@ namespace Microsoft.AppCenter.iOS.Bindings
         [Export("setLogLevel:")]
         void SetLogLevel(MSACLogLevel logLevel);
 
+        // + (void)setNetworkRequestsAllowed:(BOOL)isAllowed;
+        [Static]
+        [Export("setNetworkRequestsAllowed:")]
+        void SetNetworkRequestsAllowed(bool isAllowed);
+
+        // + (BOOL)isNetworkRequestsAllowed;
+        [Static]
+        [Export("isNetworkRequestsAllowed")]
+        bool IsNetworkRequestsAllowed();
+
         // + (void)setUserId:(NSString *)userId;
         [Static]
         [Export("setUserId:")]