+++ SDK/AppCenter/Microsoft.AppCenter.iOS/AppCenter.cs
@@ -79,6 +79,12 @@ namespace Microsoft.AppCenter
             }
         }
 
+        static bool PlatformIsNetworkRequestsAllowed
+        {
+            get => iOSAppCenter.IsNetworkRequestsAllowed();
+            set => iOSAppCenter.SetNetworkRequestsAllowed(value);
+        }
+
         static void PlatformSetUserId(string userId)
         {
             iOSAppCenter.SetUserId(userId);