+++ Tests/Microsoft.AppCenter.Test.Windows/AppCenterTest.cs
@@ -379,9 +379,9 @@ namespace Microsoft.AppCenter.Test
         public void Configure()
         {
             Assert.IsFalse(AppCenter.Configured);
-
             AppCenter.Configure("some string");
 
+            // Verify configuration settings.
             Assert.IsTrue(AppCenter.Configured);
             _channelMock.Verify(channel => channel.SetEnabled(It.IsAny<bool>()));
         }
@@ -420,6 +420,39 @@ namespace Microsoft.AppCenter.Test
             {
                 Assert.Fail();
             }
+        }
+
+        /// <summary>
+        /// Verify setting allow network requests. 
+        /// </summary>
+        [TestMethod]
+        public void TestIsNetworkRequestsAllowed()
+        {
+            AppCenter.Configure("appsecret");
+
+            // Verify setting true value when previous value is true.
+            _settingsMock.Setup(settings => settings.GetValue(AppCenter.AllowedNetworkRequestsKey, It.IsAny<bool>()))
+                .Returns(true);
+            AppCenter.IsNetworkRequestsAllowed = true;
+            _channelGroupMock.Verify(groups => groups.SetNetworkRequestAllowed(It.IsAny<bool>()), Times.Never());
+
+            // Verify setting false value when previous value is false.
+            _settingsMock.Setup(settings => settings.GetValue(AppCenter.AllowedNetworkRequestsKey, It.IsAny<bool>()))
+                .Returns(false);
+            AppCenter.IsNetworkRequestsAllowed = false;
+            _channelGroupMock.Verify(groups => groups.SetNetworkRequestAllowed(It.IsAny<bool>()), Times.Never());
+
+            // Verify setting true value when previous value is false.
+            _settingsMock.Setup(settings => settings.GetValue(AppCenter.AllowedNetworkRequestsKey, It.IsAny<bool>()))
+                .Returns(false);
+            AppCenter.IsNetworkRequestsAllowed = true;
+            _channelGroupMock.Verify(groups => groups.SetNetworkRequestAllowed(true));
+
+            // Verify setting false value when previous value is true.
+            _settingsMock.Setup(settings => settings.GetValue(AppCenter.AllowedNetworkRequestsKey, It.IsAny<bool>()))
+                .Returns(true);
+            AppCenter.IsNetworkRequestsAllowed = false;
+            _channelGroupMock.Verify(groups => groups.SetNetworkRequestAllowed(false));
         }
 
         /// <summary>