+++ Tests/Microsoft.AppCenter.Test.Windows/Ingestion/Http/RetryableTest.cs
@@ -6,6 +6,8 @@ using System.Net;
 using System.Threading.Tasks;
 using Microsoft.AppCenter.Ingestion;
 using Microsoft.AppCenter.Ingestion.Http;
+using Microsoft.AppCenter.Test.Utils;
+using Microsoft.AppCenter.Utils;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 
@@ -21,12 +23,18 @@ namespace Microsoft.AppCenter.Test.Windows.Ingestion.Http
             TimeSpan.FromSeconds(1) 
         };
         private IIngestion _retryableIngestion;
+        private readonly Mock<IApplicationSettings> _settingsMock = new Mock<IApplicationSettings>();
 
         [TestInitialize]
         public void InitializeRetryableTest()
         {
             _adapter = new Mock<IHttpNetworkAdapter>();
             _retryableIngestion = new RetryableIngestion(new IngestionHttp(_adapter.Object), Intervals);
+            AppCenter.Instance = null;
+#pragma warning disable 612
+            AppCenter.SetApplicationSettingsFactory(new MockApplicationSettingsFactory(_settingsMock));
+            _settingsMock.Setup(settings => settings.GetValue(AppCenter.AllowedNetworkRequestsKey, It.IsAny<bool>())).Returns(true);
+#pragma warning restore 612
         }
 
         /// <summary>
@@ -49,6 +57,7 @@ namespace Microsoft.AppCenter.Test.Windows.Ingestion.Http
         [TestMethod]
         public async Task RetryableIngestionRepeat1()
         {
+
             // RequestTimeout - retryable
             SetupAdapterSendResponse(HttpStatusCode.RequestTimeout, HttpStatusCode.OK);
             var start = DateTime.Now;
@@ -68,6 +77,7 @@ namespace Microsoft.AppCenter.Test.Windows.Ingestion.Http
         [TestMethod]
         public async Task RetryableIngestionRepeat3()
         {
+
             // RequestTimeout - retryable
             SetupAdapterSendResponse(HttpStatusCode.RequestTimeout, HttpStatusCode.RequestTimeout, HttpStatusCode.RequestTimeout, HttpStatusCode.OK);
             var start = DateTime.Now;
@@ -95,6 +105,7 @@ namespace Microsoft.AppCenter.Test.Windows.Ingestion.Http
         [TestMethod]
         public async Task RetryableIngestionCancel()
         {
+
             // RequestTimeout - retryable
             SetupAdapterSendResponse(HttpStatusCode.RequestTimeout);
             var start = DateTime.Now;